+++ CHANGELOG.md
@@ -1,5 +1,3 @@
-Version numbers correspond to `bower.json` version
-
 # 3.0.0 (Unpublished)
 
 ## Backward compatible changes and new features
@@ -13,6 +11,7 @@ Version numbers correspond to `bower.json` version
 
 - now throwing errors throughout the api to validate it is being used appropriately. Any properties specified on a formly field that are not explicitly part of the api will result in an error thrown. Use `data` or `templateOptions` for custom field values.
 - removing `setTemplate` and `setTemplateUrl` in favor of the new `setType` api.
+- `expressionProperties` has been simplified and improved. No longer can you specify a `data` object. Instead you specify an expression that the result of the expression will be assigned to (uses the `$parse` service).
 
 ## Internal Changes